+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Change: updated MSRV to 1.79.
 - Change: update some dependencies.
 - Change: remove unused dependencies from all packages.
+- Change(server): on rusty backend, use upstream rodio again.
 - Feat: more immediate event changes (Example: updated via mpris)
 - Feat(server): on rusty backend, enable `aiff` codec support.
 - Fix: change default config ip address to `::1` instead of `::` (any old values on windows will need to be changed manually)
@@ -15,7 +16,6 @@
 - Fix: correctly write LRC milliseconds.
 - Fix: change Lyric::adjust_offset to not work invertedly for below 10 seconds anymore.
 - Fix(tui): base "no lyrics available" message on the same value as actual parsed lyrics.
-
 
 ### [V0.9.1]
 - Released on: August 21, 2024.